backend/requirements.txt
Flask>=2.2,<3.0
flask-bcrypt
flask-cors
pymongo
PyJWT
